requirements.txt
# Core dependencies
flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.1

# PDF processing
PyMuPDF==1.24.0
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
easyocr==1.7.1

# RAG system
langchain>=0.1.0
langchain-community>=0.0.28
langchain-core>=0.1.31
langchain-ollama>=0.2.0
faiss-cpu==1.7.4